# ServerInfrastructure/Swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Open Spaces
  description: |-
    This is the API documentation for the Open Spaces backend and available routes. Open Spaces is a booking platform for reserving free rooms and hotdesks in UNSW. It is powered by Nodejs and express... [Insert Yap here]
    
    Project links:
    - [WebApp](https:/openspaces.penguinserver.net/)
    - [Github repository](https://github.com/unsw-cse-comp99-3900-24t1/capstone-project-3900f18bpookie)
    - [Figma](https://www.figma.com/design/fLKzMmwO7utmv4O1GKp7pW/Comp3900-Storyboard?node-id=0-1)
  version: 2.16.5
servers:
  - url: https://backend.openspaces.penguinserver.net/
  - url: http://localhost:5000/
...
paths:
  /user/create:
    post:
      tags:
        - user
      summary: Create user
      description: Creates a new user in the database
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
        '409':
          description: User Already Exists
        '403':
          description: Email must belong to UNSW
  /user/verify:
    post:
      tags:
        - user
      summary: Verify user
      description: Verifies user account. User will not be signed in automatically once verified.
      operationId: createUsersWithListInput
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserVerify'
      responses:
        '200':
          description: Success
        '404':
          description: User does not exists
        '403':
          description: Incorrect code
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: 'Logs user into the system using username and password'
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login (their email)
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserToken'          
            application/xml:
              schema:
                $ref: '#/components/schemas/UserToken'
        '403':
          description: Invalid username/password
        '404':
          description: User does not exist
  /user/logout:
    get:
      tags:
        - user
      summary: Logs out user's current session
      description: 'Invalidates provided session token'
      operationId: logoutUser
      parameters:
        - name: token
          in: query
          description: Session token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
        '404':
          description: Session not found
  /user/get:
    get:
      tags:
        - user
      summary: Get user by user name
      description: 'username field is optional, if username is not supplied,
      the details of the current user will be returned. User must be an admin
      to use the username field.'
      operationId: getUserByName
      parameters:
        - name: username
          in: query
          description: 'The email of the user that needs to be fetched.'
          required: false
          schema:
            type: string
        - name: token
          in: query
          description: "Logged in user's token"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad Request
        '403':
          description: User must be an admin to get other user's details
        '404':
          description: User not found
  /user/sendCode:
    get:
      tags:
        - user
      summary: Send verification code
      description: 'Sends a verification code to the user`s email'
      operationId: sendverificationCodeUser
      parameters:
        - name: username
          in: query
          description: The email of the user attempting to verify their account
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
        '403':
          description: Invalid username
  /user/delete:
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by cse admins.
      operationId: deleteUser
      parameters:
        - name: username
          in: query
          description: The email of the user that needs to be deleted.
          required: true
          schema:
            type: string
        - name: token
          in: query
          description: The signed in user's token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
        '403':
          description: User must be an admin to delete other users
        '404':
          description: User does not exist
  /Booking/create:
    post:
      tags:
        - bookings
      summary: Create a booking
      description: Creates a new booking for a room or hot desk
      requestBody:
        description: Booking details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBooking'
      responses:
        '200':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /booking/edit:
    post:
      tags:
        - bookings
      summary: Create a booking
      description: Creates a new booking for a room or hot desk
      requestBody:
        description: Booking details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditBooking'
      responses:
        '200':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /booking/override:
    post:
      tags:
        - bookings
      summary: Create a booking
      description: Creates a new booking for a room or hot desk
      requestBody:
        description: Booking details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OverrideBooking'
      responses:
        '200':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/get/{bookingId}:
    get:
      tags:
        - bookings
      summary: Finds a booking by ID
      description: HAS NOT BEEN IMPLEMENTED YET!!!
      operationId: getOrderById
      parameters:
        - name: bookingId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/feedback'          
            application/xml:
              schema:
                $ref: '#/components/schemas/feedback'
        '400':
          description: Invalid ID supplied
        '404':
          description: Booking not found
  /bookings/availableSpaces/{spaceID}:
    get:
      tags:
        - bookings
      summary: Get available spaces for a given space and time range
      description: Retrieve available time slots for the given space ID, start time, and end time.
      parameters:
        - name: spaceID
          in: path
          required: true
          description: The ID of the space
          schema:
            type: integer
        - name: startTime
          in: query
          required: true
          description: The start time of the desired booking slot
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          description: The end time of the desired booking slot
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimeSlot'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error message"
  /bookings/availableSpaces/{floorplanID}:
    get:
      tags:
        - bookings
      summary: Get available spaces for a given floor plan and time range
      description: Retrieve available rooms or desks for the given floor plan ID, start time, and end time.
      parameters:
        - name: floorplanID
          in: path
          required: true
          description: The ID of the floor plan
          schema:
            type: integer
        - name: startTime
          in: query
          required: true
          description: The start time of the desired booking slot
          schema:
            type: string
            format: date-time
        - name: endTime
          in: query
          required: true
          description: The end time of the desired booking slot
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Space'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error message"
  /bookings/list/{spaceID}:
    get:
      tags:
        - bookings
      summary: List bookings for a space
      description: Retrieve all active bookings for a given space ID. Only accessible by admins.
      parameters:
        - name: spaceID
          in: path
          required: true
          description: The ID of the space
          schema:
            type: integer
        - name: token
          in: query
          required: true
          description: The admin's token
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required for this operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/getUserAll:
    get:
      tags:
        - bookings
      summary: Finds a booking by using token
      description: gets all bookings active
      operationId: getbookinguserall
      parameters:
        - name: token
          in: query
          description: The signed in user's token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
        '400':
          description: Invalid ID supplied
        '404':
          description: Booking not found
  /Bookings/delete:
    delete:
      tags:
        - bookings
      summary: remove booking
      description: This can only be done by cse admins, or person booking
      operationId: removeBooking
      parameters:
        - name: BookingID
          in: query
          description: The BookingID of the booking that needs to be deleted.
          required: true
          schema:
            type: integer
        - name: token
          in: query
          description: The signed in user's token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
        '403':
          description: User must be an admin to remove booking
        '404':
          description: Booking does not exist
...
  /space/checkIn:
    post:
      tags:
        - spaces
      summary: Create space
      description: Creates a new space in the database
      operationId: spaceCheckIn
      requestBody:
        description: created upject
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpaceCheckIn'
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
        '409':
          description: User Already Exists
        '403':
          description: Email must belong to UNSW    
...
components:
  schemas:
...
    UserCreate:
      type: object
      properties:
        firstName:
          type: string
          example: 'John'
        lastName:
          type: string
          example: 'James'
        email:
          type: string
          example: 'john@unsw.edu.au'
        password:
          type: string
          example: '12345'
      xml:
        name: usercreate
    UserVerify:
      type: object
      properties:
        username:
          type: string
          example: 'john@unsw.edu.au'
        verificationCode:
          type: integer
          example: 123456
      xml:
        name: userverify
...
    UserToken:
      type: object
      properties:
        token:
          description: jwt token
          type: string
          example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c'
...